requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

gunicorn==23.0.0  # https://github.com/benoitc/gunicorn
psycopg[c]==3.2.9  # https://github.com/psycopg/psycopg

# Django
# ------------------------------------------------------------------------------
django-anymail==13.0  # https://github.com/anymail/django-anymail
channels>=4.0   #
channels-redis>=4.0  #
daphne>=4.0   #
djangorestframework-simplejwt>=5.5.0  #
djangorestframework-simplejwt[blacklist]
django-filter>=23.2.0
